FirstShop/requirements.txt
Django==2.2.4
django-jet==1.0.8
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
django-filter==2.2.0
django-cors-headers 3.2.1
